Add Some Validation Class

+++ Accounting/Accounting/DataLayer/Entities/ProductSold.cs
@@ -18,7 +18,7 @@ namespace Accounting.DataLayer.Entities
         public string Invoice { get; set; }
         public Product ProductId { get; set; }
         public int Quantity { get; set; }
-        public int Price { get; set; }
-        public int TotalAmount { get; set; }
+        public decimal Price { get; set; }
+        public decimal TotalAmount { get; set; }
     }
 }
